+++ src/test/java/bio/terra/controller/StudyTest.java
@@ -4,25 +4,16 @@ import bio.terra.category.Unit;
 import bio.terra.controller.exception.ApiException;
 import bio.terra.dao.StudyDao;
 import bio.terra.dao.exception.StudyNotFoundException;
+import bio.terra.fixtures.FlightStates;
 import bio.terra.flight.study.create.StudyCreateFlight;
 import bio.terra.metadata.Study;
-import bio.terra.model.AssetModel;
-import bio.terra.model.AssetTableModel;
-import bio.terra.model.ColumnModel;
-import bio.terra.model.RelationshipModel;
-import bio.terra.model.RelationshipTermModel;
 import bio.terra.model.StudyJsonConversion;
 import bio.terra.model.StudyModel;
-import bio.terra.model.StudyRequestModel;
-import bio.terra.model.StudySpecificationModel;
-import bio.terra.model.TableModel;
 import bio.terra.stairway.FlightMap;
 import bio.terra.stairway.FlightState;
 import bio.terra.stairway.Stairway;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.io.IOUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
@@ -36,10 +27,9 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 import java.time.Instant;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.UUID;
 
+import static bio.terra.fixtures.StudyFixtures.studyRequest;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
 import static org.mockito.ArgumentMatchers.eq;
@@ -49,9 +39,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-
-import bio.terra.fixtures.FlightStates;
 
 
 @RunWith(SpringRunner.class)
@@ -69,82 +56,8 @@ public class StudyTest {
     @Autowired
     private ObjectMapper objectMapper;
 
-    private TableModel participantTable;
-    private TableModel sampleTable;
-    private RelationshipTermModel participantTerm;
-    private RelationshipTermModel sampleTerm;
-    private RelationshipModel participantSampleRelationship;
-    private AssetTableModel assetParticipantTable;
-    private AssetTableModel assetSampleTable;
-    private AssetModel asset;
-    private StudySpecificationModel schema;
-    private StudyRequestModel studyRequest;
 
     private static final String testFlightId = "test-flight-id";
-
-    @Before
-    public void setup() {
-        participantTable = new TableModel()
-                .name("participant")
-                .columns(Arrays.asList(
-                        new ColumnModel().name("id").datatype("string"),
-                        new ColumnModel().name("age").datatype("number")));
-
-        sampleTable = new TableModel()
-                .name("sample")
-                .columns(Arrays.asList(
-                        new ColumnModel().name("id").datatype("string"),
-                        new ColumnModel().name("participant_id").datatype("string"),
-                        new ColumnModel().name("date_collected").datatype("date")));
-
-        participantTerm = new RelationshipTermModel()
-                .table("participant")
-                .column("id")
-                .cardinality(RelationshipTermModel.CardinalityEnum.ONE);
-
-        sampleTerm = new RelationshipTermModel()
-                .table("sample")
-                .column("participant_id")
-                .cardinality(RelationshipTermModel.CardinalityEnum.MANY);
-
-        participantSampleRelationship = new RelationshipModel()
-                .name("participant_sample")
-                .from(participantTerm)
-                .to(sampleTerm);
-
-        assetParticipantTable = new AssetTableModel()
-                .name("participant")
-                .columns(Collections.emptyList());
-
-        assetSampleTable = new AssetTableModel()
-                .name("sample")
-                .columns(Arrays.asList("participant_id", "date_collected"));
-
-        asset = new AssetModel()
-                .name("Sample")
-                .rootTable("sample")
-                .rootColumn("participant_id")
-                .tables(Arrays.asList(assetParticipantTable, assetSampleTable))
-                .follow(Collections.singletonList("participant_sample"));
-
-        schema = new StudySpecificationModel()
-                .tables(Arrays.asList(participantTable, sampleTable))
-                .relationships(Collections.singletonList(participantSampleRelationship))
-                .assets(Collections.singletonList(asset));
-
-        studyRequest = new StudyRequestModel()
-                .name("Minimal")
-                .description("This is a sample study definition")
-                .schema(schema);
-
-    }
-
-    private void expectBadStudyCreateRequest(StudyRequestModel studyRequest) throws Exception {
-        mvc.perform(post("/api/repository/v1/studies")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(studyRequest)))
-                .andExpect(status().is4xxClientError());
-    }
 
     @Test
     public void testMinimalCreate() throws Exception {
@@ -187,14 +100,6 @@ public class StudyTest {
     }
 
     @Test
-    public void testInvalidStudyRequest() throws Exception {
-        mvc.perform(post("/api/repository/v1/studies")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content("{}"))
-                .andExpect(status().is4xxClientError());
-    }
-
-    @Test
     public void testFlightError() throws Exception {
         when(stairway.submit(eq(StudyCreateFlight.class), isA(FlightMap.class)))
                 .thenThrow(ApiException.class);
@@ -202,147 +107,6 @@ public class StudyTest {
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(studyRequest)))
                 .andExpect(status().isInternalServerError());
-    }
-
-    @Test
-    public void testDuplicateTableNames() throws Exception {
-        ColumnModel column = new ColumnModel().name("id").datatype("string");
-        TableModel table = new TableModel()
-                .name("duplicate")
-                .columns(Collections.singletonList(column));
-
-        studyRequest.getSchema().tables(Arrays.asList(table, table));
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testDuplicateColumnNames() throws Exception {
-        ColumnModel column = new ColumnModel().name("id").datatype("string");
-        TableModel table = new TableModel()
-                .name("table")
-                .columns(Arrays.asList(column, column));
-
-        studyRequest.getSchema().tables(Collections.singletonList(table));
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testMissingAssets() throws Exception {
-        studyRequest.getSchema().assets(null);
-        expectBadStudyCreateRequest(studyRequest);
-
-        studyRequest.getSchema().assets(Collections.emptyList());
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testDuplicateAssetNames() throws Exception {
-        studyRequest.getSchema().assets(Arrays.asList(asset, asset));
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testDuplicateRelationshipNames() throws Exception {
-        studyRequest.getSchema()
-                .relationships(Arrays.asList(participantSampleRelationship, participantSampleRelationship));
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testInvalidAssetTable() throws Exception {
-        AssetTableModel invalidAssetTable = new AssetTableModel()
-                .name("mismatched_table_name")
-                .columns(Collections.emptyList());
-
-        AssetModel asset = new AssetModel()
-                .name("bad_asset")
-                .rootTable("mismatched_table_name")
-                .tables(Collections.singletonList(invalidAssetTable))
-                .follow(Collections.singletonList("participant_sample"));
-
-        studyRequest.getSchema().assets(Collections.singletonList(asset));
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testInvalidAssetTableColumn() throws Exception {
-        // participant is a valid table but date_collected is in the sample table
-        AssetTableModel invalidAssetTable = new AssetTableModel()
-                .name("participant")
-                .columns(Collections.singletonList("date_collected"));
-
-        AssetModel asset = new AssetModel()
-                .name("mismatched")
-                .rootTable("participant")
-                .rootColumn("date_collected")
-                .tables(Collections.singletonList(invalidAssetTable))
-                .follow(Collections.singletonList("participant_sample"));
-
-        studyRequest.getSchema().assets(Collections.singletonList(asset));
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testInvalidFollowsRelationship() throws Exception {
-        AssetModel asset = new AssetModel()
-                .name("bad_follows")
-                .tables(Arrays.asList(assetSampleTable, assetParticipantTable))
-                .follow(Collections.singletonList("missing"));
-
-        studyRequest.getSchema().assets(Collections.singletonList(asset));
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testInvalidRelationshipTermTableColumn() throws Exception {
-        // participant_id is part of the sample table, not participant
-        RelationshipTermModel mismatchedTerm = new RelationshipTermModel()
-               .table("participant")
-               .column("participant_id")
-               .cardinality(RelationshipTermModel.CardinalityEnum.ONE);
-
-        RelationshipModel mismatchedRelationship = new RelationshipModel()
-                .name("participant_sample")
-                .from(mismatchedTerm)
-                .to(sampleTerm);
-
-        studyRequest.getSchema().relationships(Collections.singletonList(mismatchedRelationship));
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-
-    @Test
-    public void testNoRootTable() throws Exception {
-        AssetModel noRoot = new AssetModel()
-                .name("bad")
-                // In the fixtures, the participant asset table has isRoot set to false.
-                .tables(Collections.singletonList(assetParticipantTable))
-                .follow(Collections.singletonList("participant_sample"));
-        studyRequest.getSchema().assets(Collections.singletonList(noRoot));
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testStudyNameInvalid() throws Exception {
-        studyRequest.name("no spaces");
-        expectBadStudyCreateRequest(studyRequest);
-
-        studyRequest.name("no-dashes");
-        expectBadStudyCreateRequest(studyRequest);
-
-        studyRequest.name("");
-        expectBadStudyCreateRequest(studyRequest);
-
-        // Make a 64 character string, it should be considered too long by the validation.
-        String tooLong = StringUtils.repeat("a", 64);
-        studyRequest.name(tooLong);
-        expectBadStudyCreateRequest(studyRequest);
-    }
-
-    @Test
-    public void testStudyNameMissing() throws Exception {
-        studyRequest.name(null);
-        expectBadStudyCreateRequest(studyRequest);
     }
 
     @MockBean
